Tf_MNB_Fakenews-main/Tf_MNB_Fakenews-main/requirements.txt
beautifulsoup4==4.11.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
cssselect==1.1.0
cycler==0.11.0
docopt==0.6.2
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.7.1
Flask==2.1.1
Flask-Cors==3.0.10
Flask-WTF==1.0.1
fonttools==4.33.3
gunicorn==20.1.0
idna==3.3
imageio==2.19.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
jieba3k==0.35.1
Jinja2==3.1.1
joblib==1.0.1
kiwisolver==1.4.2
lime==0.2.0.1
lxml==4.9.0
MarkupSafe==2.1.1
matplotlib==3.5.2
networkx==2.6.3
newspaper3k==0.2.8
nltk==3.6.2
numpy==1.21.6
packaging==21.3
pandas==1.3.5
Pillow==9.1.1
pipreqs==0.4.11
pyparsing==3.0.9
pytz==2022.1
PyWavelets==1.3.0
pywin32==303; platform_system=="Windows"
PyYAML==6.0
regex==2021.4.4
requests==2.27.1
requests-file==1.5.1
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.7.3
sgmllib3k==1.0.0
six==1.16.0
sklearn==0.0
soupsieve==2.3.2.post1
threadpoolctl==3.1.0
tifffile==2021.11.2
tinysegmenter==0.3
tldextract==3.3.0
tqdm==4.60.0
urllib3==1.26.9
vaderSentiment==3.3.2
wcwidth==0.2.5
Werkzeug==2.1.1
WTForms==3.0.1
yarg==0.1.9
zipp==3.8.0
